encanta/apps/api/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
python-multipart==0.0.6
httpx>=0.23.0,<0.24.0
openai>=1.6.1,<2.0.0
supabase==1.0.3
pinecone-client==2.2.2
redis==4.6.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
pytest==7.4.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
langchain==0.0.267
langchain-openai==0.0.2
tiktoken>=0.5.2,<0.6.0
tenacity==8.2.3
stripe==11.6.0 